Fixed salt id type

+++ authentication/authentication-common/src/main/java/io/mosip/authentication/common/service/impl/idevent/CredentialStoreService.java
@@ -402,7 +402,7 @@ public class CredentialStoreService {
 	 * @param salt   the salt
 	 */
 	private void saveSalt(String modulo, String salt) {
-		Long saltModulo = Long.valueOf(modulo);
+		Integer saltModulo = Integer.valueOf(modulo);
 		if (!uinHashSaltRepo.existsById(saltModulo)) {
 			UinHashSalt saltEntity = new UinHashSalt();
 			saltEntity.setId(saltModulo);
